backend/requirements/dev.txt
-r base.txt

# Development tools
django-debug-toolbar==4.2.0
django-debug-toolbar-template-profiler==2.1.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0

# Code quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
pre-commit==3.5.0
mypy==1.7.1
django-stubs==4.2.7

# Development server
watchdog==3.0.0